Marketplace/src/storage/customer_db/requirements.txt
spyne==2.14.0
lxml
connexion >= 2.6.0
connexion[swagger-ui] >= 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2
fuzzywuzzy==0.18.0
grpcio-tools==1.51.1
suds
flask
jsonpickle